Extract entry provider query methods to trait
