+++ src/Blog.php
@@ -8,12 +8,14 @@ use Bjuppa\LaravelBlog\Contracts\BlogEntryProvider;
 use Bjuppa\LaravelBlog\Support\Author;
 use Bjuppa\LaravelBlog\Support\ProvidesBladeViews;
 use Bjuppa\LaravelBlog\Support\PrefixesRouteNames;
-use Carbon\Carbon;
+use Bjuppa\LaravelBlog\Support\QueriesEntryProvider;
 use Illuminate\Support\Collection;
 
 class Blog implements BlogContract
 {
-    use ProvidesBladeViews, PrefixesRouteNames;
+    use ProvidesBladeViews;
+    use PrefixesRouteNames;
+    use QueriesEntryProvider;
 
     /**
      * Blog identifier
@@ -185,28 +187,6 @@ class Blog implements BlogContract
     }
 
     /**
-     * Get an entry instance from a slug
-     *
-     * @param string $slug
-     * @return BlogEntry|null
-     */
-    public function findEntry(string $slug): ?BlogEntry
-    {
-        return $this->getEntryProvider()->findBySlug($slug);
-    }
-
-    /**
-     * Get the newest entries of the blog
-     *
-     * @param int|null $limit Desired number of entries unless you want the blog's default
-     * @return Collection
-     */
-    public function latestEntries(int $limit = null): Collection
-    {
-        return $this->getEntryProvider()->latest($limit ?? $this->getLatestEntriesLimit());
-    }
-
-    /**
      * Get the number of default entries to show
      *
      * @return int
@@ -302,15 +282,6 @@ class Blog implements BlogContract
         $this->title = $title;
 
         return $this;
-    }
-
-    /**
-     * Get the last updated timestamp for the entire blog
-     * @return Carbon
-     */
-    public function getUpdated(): Carbon
-    {
-        return $this->getEntryProvider()->getUpdated();
     }
 
     /**
